replaced stop button by mousereleased event, and renamed fxml and package
