+++ src/SolArduino/Main.java
@@ -1,4 +1,4 @@
-package sample;
+package SolArduino;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -10,7 +10,7 @@ public class Main extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception{
-        Parent root = FXMLLoader.load(getClass().getResource("sample.fxml"));
+        Parent root = FXMLLoader.load(getClass().getResource("interface.fxml"));
         primaryStage.setTitle("SolArduino");
         primaryStage.setScene(new Scene(root, 500, 300));
         primaryStage.show();
